+++ src/internal/types/index.ts
@@ -1,3 +1,3 @@
 export type { MarquaData, MarquaTable } from './data';
 export type { FileOptions, DirOptions, ParserTypes } from './option';
-export type { FrontMatter, HydrateFn } from './parser';
+export type { FrontMatter, Hydrate } from './parser';