+++ src/internal/types/option.ts
@@ -1,8 +1,8 @@
-declare const UnSym1: unique symbol;
-declare const UnSym2: unique symbol;
+declare const UniqueInput: unique symbol;
+declare const UniqueOutput: unique symbol;
 export type ParserTypes<Input, Output> = {
-	[UnSym1]: Input;
-	[UnSym2]: Output;
+	[UniqueInput]: Input;
+	[UniqueOutput]: Output;
 };
 
 export type FileOptions = {
@@ -11,8 +11,12 @@ export type FileOptions = {
 	exclude?: Array<string>;
 };
 
-export type DirOptions = FileOptions & {
+export type DirOptions<Output extends object = {}> = FileOptions & {
 	entry: string;
 	recurse?: boolean;
 	extensions?: Array<string>;
+	sort?(
+		x: keyof Output extends never ? Record<string, any> : Output,
+		y: keyof Output extends never ? Record<string, any> : Output
+	): number;
 };