+++ test/typings/core/default-dir/recurse.test-d.ts
@@ -21,8 +21,10 @@ expectType<Array<Generic>>(
 	traverse(options, ({ frontMatter, content, breadcrumb }) => {
 		expectType<Default>(frontMatter);
 
-		expectType<Array<ts.MarquaData>>(content);
+		expectType</*Array< ts.MarquaData >*/ string>(content);
 
 		expectType<Array<string>>(breadcrumb);
+
+		return undefined;
 	})
 );